requirements.txt
flask
pymongo
kafka-python
python-dotenv
